+++ patches/src/main/kotlin/app/revanced/patches/shared/misc/spoof/SpoofVideoStreamsPatch.kt
@@ -10,8 +10,10 @@ import app.revanced.patcher.patch.BytecodePatchContext
 import app.revanced.patcher.patch.bytecodePatch
 import app.revanced.patcher.util.proxy.mutableTypes.MutableMethod.Companion.toMutable
 import app.revanced.patches.all.misc.resources.addResourcesPatch
+import app.revanced.util.findInstructionIndicesReversedOrThrow
 import app.revanced.util.getReference
 import app.revanced.util.indexOfFirstInstructionOrThrow
+import app.revanced.util.insertFeatureFlagBooleanOverride
 import app.revanced.util.returnEarly
 import com.android.tools.smali.dexlib2.AccessFlags
 import com.android.tools.smali.dexlib2.Opcode
@@ -206,6 +208,34 @@ fun spoofVideoStreamsPatch(
                     """,
             )
         }
+
+        // endregion
+
+        // region Append spoof info.
+
+        nerdsStatsVideoFormatBuilderFingerprint.method.apply {
+            findInstructionIndicesReversedOrThrow(Opcode.RETURN_OBJECT).forEach { index ->
+                val register = getInstruction<OneRegisterInstruction>(index).registerA
+
+                addInstructions(
+                    index,
+                    """
+                        invoke-static { v$register }, $EXTENSION_CLASS_DESCRIPTOR->appendSpoofedClient(Ljava/lang/String;)Ljava/lang/String;
+                        move-result-object v$register
+                    """
+                )
+            }
+        }
+
+        // endregion
+
+        // region Fix iOS livestream current time.
+
+        hlsCurrentTimeFingerprint.method.insertFeatureFlagBooleanOverride(
+            HLS_CURRENT_TIME_FEATURE_FLAG,
+            "$EXTENSION_CLASS_DESCRIPTOR->fixHLSCurrentTime(Z)Z"
+        )
+
         // endregion
 
         executeBlock()