+++ patches/src/main/kotlin/app/revanced/patches/youtube/misc/spoof/SpoofVideoStreamsPatch.kt
@@ -38,10 +38,20 @@ val spoofVideoStreamsPatch = spoofVideoStreamsPatch({
             preferences = setOf(
                 SwitchPreference("revanced_spoof_video_streams"),
                 ListPreference(
+                    "revanced_spoof_video_streams_client_type",
+                    summaryKey = null,
+                ),
+                NonInteractivePreference(
+                    // Requires a key and title but the actual text is chosen at runtime.
+                    key = "revanced_spoof_video_streams_about_android",
+                    tag = "app.revanced.extension.youtube.settings.preference.SpoofStreamingDataSideEffectsPreference"
+                ),
+                ListPreference(
                     "revanced_spoof_video_streams_language",
                     summaryKey = null
                 ),
-                NonInteractivePreference("revanced_spoof_video_streams_about")
+                SwitchPreference("revanced_spoof_video_streams_ios_force_avc"),
+                SwitchPreference("revanced_spoof_streaming_data_stats_for_nerds"),
             ),
         ),
     )